+++ NEWS.md
@@ -24,6 +24,7 @@
 #### Changed
 * Pretty printing for tibbles removed as it is not really the scope of this package
 * Improved speed of key antibiotics comparison for determining first isolates
+* Column names for the `key_antibiotics` function are now generic: 6 for broadspectrum ABs, 6 for Gram-positive specific and 6 for Gram-negative specific ABs
 * Printing of class `mic` now shows all MIC values
 * `%like%` now supports multiple patterns
 * Frequency tables are now actual `data.frame`s with altered console printing to make it look like a frequency table. Because of this, the parameter `toConsole` is not longer needed.